+++ src/main/java/com/hackerramp/subscription/controller/OrdersController.java
@@ -1,7 +1,7 @@
 package com.hackerramp.subscription.controller;
 
-import com.hackerramp.subscription.services.entities.OrderDetailsResponse;
-import com.hackerramp.subscription.services.entities.OrderRequest;
+import com.hackerramp.subscription.services.beans.OrderDetailsResponse;
+import com.hackerramp.subscription.services.beans.OrderRequest;
 import com.sun.istack.internal.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
