+++ src/main/java/com/hackerramp/subscription/services/SubscriptionService.java
@@ -1,0 +1,7 @@
+package com.hackerramp.subscription.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class SubscriptionService {
+}
