+++ src/main/java/com/hackerramp/subscription/services/beans/SubscriptionRequest.java
@@ -1,21 +1,17 @@
-package com.hackerramp.subscription.services.entities;
+package com.hackerramp.subscription.services.beans;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 import lombok.ToString;
 
-import java.util.Date;
-
 @Data
 @ToString
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class SubscriptionEntry {
+public class SubscriptionRequest {
     private Long subId;
+    private String uidx;
     private String productId;
-    private Date orderPlacedDate;
-    private Date nextOrderTriggerDate;
     private Long intervalInSec;
     private String address;
     private String paymentMode;
-    private String deliveriesLeft;
 }
